+++ iolanta/errors.py
@@ -1,7 +1,10 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
+from typing import Optional
 
 from documented import DocumentedError
 from rdflib.term import Node
+
+from iolanta.models import LDContext
 
 
 @dataclass
@@ -26,3 +29,28 @@ class InsufficientDataForRender(DocumentedError):
             )
 
         return hopeless
+
+
+@dataclass
+class UnresolvedIRI(DocumentedError):
+    """
+    An unresolved IRI found.
+
+        IRI: {self.iri}
+        file: {self.file}
+        prefix: {self.prefix}
+
+    Perhaps you forgot to import appropriate context? For example:
+
+    ```yaml
+    "@context":
+        - {self.prefix}: https://example.com/{self.prefix}/
+    ```
+
+    Context: {self.context}
+    """
+
+    iri: str
+    prefix: str
+    file: Optional[str] = None    # noqa: WPS110
+    context: Optional[LDContext] = None
